create SearchService
