+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/controllers/pub/SearchController.java
@@ -2,11 +2,11 @@ package net.osmand.server.controllers.pub;
 
 import com.google.gson.Gson;
 import net.osmand.data.Amenity;
-import net.osmand.data.QuadRect;
 import net.osmand.data.Street;
 import net.osmand.search.core.ObjectType;
 import net.osmand.search.core.SearchResult;
 import net.osmand.server.api.services.OsmAndMapsService;
+import net.osmand.server.api.services.SearchService;
 import net.osmand.util.Algorithms;
 import net.osmand.util.MapUtils;
 import org.apache.commons.logging.Log;
@@ -32,13 +32,16 @@ public class SearchController {
     @Autowired
     OsmAndMapsService osmAndMapsService;
     
+    @Autowired
+    SearchService searchService;
+    
     @RequestMapping(path = "/search", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<?> search(@RequestParam double lat, @RequestParam double lon, @RequestParam String search) throws IOException, InterruptedException {
+    public ResponseEntity<String> search(@RequestParam double lat, @RequestParam double lon, @RequestParam String search) throws IOException {
         if (!osmAndMapsService.validateAndInitConfig()) {
             return osmAndMapsService.errorConfig();
         }
         try {
-            List<SearchResult> res = osmAndMapsService.search(lat, lon, search);
+            List<SearchResult> res = searchService.search(lat, lon, search);
             List<RoutingController.Feature> features = new ArrayList<>();
             int pos = 0;
             int posLoc = 0;
@@ -54,10 +57,9 @@ public class SearchController {
                             typeString += " " + (int) (sr.distRelatedObjectName / 1000.f) + " km";
                         }
                     }
-                    if (sr.objectType == ObjectType.HOUSE) {
-                        if (sr.relatedObject instanceof Street) {
+                    if (sr.objectType == ObjectType.HOUSE && (sr.relatedObject instanceof Street)) {
                             typeString += " " + ((Street) sr.relatedObject).getCity().getName();
-                        }
+                        
                     }
                     if (sr.objectType == ObjectType.LOCATION) {
                         typeString += " " + osmAndMapsService.getOsmandRegions().getCountryName(sr.location);
@@ -74,8 +76,8 @@ public class SearchController {
                             prop("index", pos).prop("locindex", posLoc).prop("distance", loc).prop("type", sr.objectType));
                 }
             }
-            return ResponseEntity.ok(gson.toJson(new RoutingController.FeatureCollection(features.toArray(new RoutingController.Feature[features.size()]))));
-        } catch (IOException | InterruptedException | RuntimeException e) {
+            return ResponseEntity.ok(gson.toJson(new RoutingController.FeatureCollection(features.toArray(new RoutingController.Feature[0]))));
+        } catch (IOException | RuntimeException e) {
             LOGGER.error(e.getMessage(), e);
             throw e;
         }
@@ -88,7 +90,7 @@ public class SearchController {
                                             @RequestParam double lon,
                                             @RequestParam int zoom) throws IOException {
         List<String> categories = (List<String>) data.get("categories");
-        OsmAndMapsService.PoiSearchResult poiSearchResult = osmAndMapsService.searchPoi(lat, lon, categories, data, zoom);
+        SearchService.PoiSearchResult poiSearchResult = searchService.searchPoi(lat, lon, categories, data, zoom);
         if (poiSearchResult != null) {
             return ResponseEntity.ok(gson.toJson(poiSearchResult));
         } else {
@@ -99,7 +101,7 @@ public class SearchController {
     @GetMapping(path = {"/get-poi-categories"}, produces = "application/json")
     @ResponseBody
     public ResponseEntity<String> getPoiCategories() {
-        Map<String, List<String>> categoriesNames = osmAndMapsService.searchPoiCategories();
+        Map<String, List<String>> categoriesNames = searchService.searchPoiCategories();
         if (categoriesNames != null) {
             return ResponseEntity.ok(gson.toJson(categoriesNames));
         } else {
@@ -110,7 +112,7 @@ public class SearchController {
     @GetMapping(path = {"/get-top-filters"}, produces = "application/json")
     @ResponseBody
     public ResponseEntity<String> getTopFilters() {
-        List<String> filters = osmAndMapsService.getTopFilters();
+        List<String> filters = searchService.getTopFilters();
         if (filters != null) {
             return ResponseEntity.ok(gson.toJson(filters));
         } else {
@@ -121,7 +123,7 @@ public class SearchController {
     @GetMapping(path = {"/search-poi-categories"}, produces = "application/json")
     @ResponseBody
     public ResponseEntity<String> searchPoiCategories(@RequestParam String search) throws IOException {
-        Map<String, Map<String, String>> res = osmAndMapsService.searchPoiCategories(search);
+        Map<String, Map<String, String>> res = searchService.searchPoiCategories(search);
         if (!res.isEmpty()) {
             return ResponseEntity.ok(gson.toJson(res));
         } else {
